requirements.txt
Django==3.0.5
djangorestframework==3.11.0
Pillow==7.1.1
python-dotenv==0.13.0